+++ Demo/Demo/SwiftUIComponents/VaultViews/SetupTokenResultView.swift
@@ -1,0 +1,44 @@
+import SwiftUI
+
+struct SetupTokenResultView: View {
+
+    @ObservedObject var vaultViewModel: VaultViewModel
+
+    var body: some View {
+        switch vaultViewModel.state.setupTokenResponse {
+        case .idle, .loading:
+            EmptyView()
+        case .loaded(let setupTokenResponse):
+            getSuccessView(setupTokenResponse: setupTokenResponse)
+        case .error(let errorMessage):
+            ErrorView(errorMessage: errorMessage)
+        }
+    }
+
+    func getSuccessView(setupTokenResponse: SetUpTokenResponse) -> some View {
+        VStack(spacing: 16) {
+            HStack {
+                Text("Setup Token")
+                    .font(.system(size: 20))
+                Spacer()
+            }
+            LeadingText("ID", weight: .bold)
+            LeadingText("\(setupTokenResponse.id)")
+            LeadingText("Customer ID", weight: .bold)
+            LeadingText("\(setupTokenResponse.customer?.id ?? "")")
+            LeadingText("Status", weight: .bold)
+            LeadingText("\(setupTokenResponse.status)")
+            if let url = setupTokenResponse.paypalURL {
+                LeadingText("PayPalURL", weight: .bold)
+                LeadingText("\(url)")
+            }
+        }
+        .frame(maxWidth: .infinity)
+        .padding()
+        .background(
+            RoundedRectangle(cornerRadius: 10)
+                .stroke(.gray, lineWidth: 2)
+                .padding(5)
+        )
+    }
+}